renaming common into exceptions and other refactoring
