+++ hivpy.conf
@@ -1,7 +1,7 @@
 [EXPERIMENT]
 POPULATION = 1000
 START_YEAR = 1989
-END_YEAR = 2071
+END_YEAR = 2021
 TIME_INTERVAL_DAYS = 90
 TIME_INTERVAL_MONTHS = 3
 NUMBER_MODEL_RUNS = 100
@@ -9,6 +9,7 @@ NUMBER_MODEL_RUNS = 100
 [GENERAL]
 OUTPUT_DIRECTORY = /Users/peterschmidt/tmp
 LOGOUTPUT_NAME = hivpy.log
-LOG_LEVEL = WARNING #DEBUG, INFO, WARNING, ERROR, CRITICAL - default is WARNING
+LOG_LEVEL = INFO
+
