+++ src/hivpy/__init__.py
@@ -1,3 +1,3 @@
-from .common import SimulationException
+from .exceptions import SimulationException
 from .experiment import run_experiment
 from .simulation import run_simulation, SimulationConfig
