+++ src/hivpy/exceptions.py
